+++ Demos/09-Testing/Testing/src/app/demos/samples/component-injection/component-injection.component.html
@@ -1,4 +1,4 @@
-<app-markdown-renderer [md]="'comp-injection'"></app-markdown-renderer>
+<!-- <app-markdown-renderer [md]="'comp-injection'"></app-markdown-renderer> -->
 
 <mat-card>
   <mat-card-header>
